Move test_stop_running into test_job_queue
The test has nothing to do with EnkfSimulationRunner.
